# api/openapi.yaml
openapi: 3.1.0
info:
  title: OnboardMePlease API
  version: 0.4.0
  description: >-
    Capture, resumable synthesis, searchable OKF knowledge and source-grounded questions are implemented.
    Deletion and separate snapshot creation remain planned; those routes are marked below.
servers:
  - url: http://127.0.0.1:8765
    description: Loopback default for the local single-user service.
paths:
  /v1/repositories:
    get:
      operationId: listRepositories
      summary: List recent public GitHub snapshots
      responses:
        '200': { description: Recent repositories }
    post:
      operationId: addRepository
      summary: Validate input and queue an immutable snapshot/index job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../contracts/repository.schema.json#/$defs/RepositoryInput'
      responses:
        '202':
          description: Repository accepted for analysis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAccepted'
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
...
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}:
    get:
      operationId: getSnapshot
      summary: Read snapshot identity and state
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      responses:
        '200':
          description: Snapshot metadata
          content:
            application/json:
              schema: { $ref: '../contracts/repository.schema.json#/$defs/Snapshot' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/coverage:
    get:
      operationId: getCoverage
      summary: Read inventory status counts
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      responses:
        '200':
          description: Snapshot coverage
          content:
            application/json:
              schema: { $ref: '../contracts/repository.schema.json#/$defs/Coverage' }
        '403': { $ref: '#/components/responses/Problem' }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/index:
    post:
      operationId: indexSnapshot
      summary: Reindex an existing ready GitHub snapshot
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      responses:
        '202': { description: Durable indexing job queued }
        '404': { $ref: '#/components/responses/Problem' }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/search:
    get:
      operationId: searchEvidence
      summary: Scoped exact/lexical or explicitly selected semantic retrieval
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 200 }
        - name: mode
          in: query
          schema: { enum: [lexical, semantic] }
        - name: provider
          in: query
          schema: { enum: [local, cloud] }
      responses:
        '200': { description: Snapshot-scoped source chunks with ranges }
        '422': { $ref: '#/components/responses/Problem' }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/capabilities:
    get:
      operationId: getCapabilities
      summary: Per-artifact analysis capabilities and gaps
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      responses:
        '200': { description: Artifact capability records }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/symbols:
    get:
      operationId: getSymbols
      summary: Snapshot-scoped extracted or SCIP-imported symbols
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      responses:
        '200': { description: Symbols and source ranges }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/graph:
    get:
      operationId: getGraph
      summary: Bounded typed relation traversal from one symbol
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
        - name: symbol_id
          in: query
          required: true
          schema: { type: string }
      responses:
        '200': { description: Nodes, edges, limits, and truncation flag }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/scip:
    post:
      operationId: importSCIP
      summary: Import matching SCIP semantic evidence
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
      responses:
        '200': { description: Imported and skipped document counts }
        '422': { $ref: '#/components/responses/Problem' }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/embeddings:
    post:
      operationId: queueEmbeddings
      summary: Explicitly queue local or opted-in cloud embeddings
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [provider]
              properties:
                provider: { enum: [local, cloud] }
      responses:
        '202': { description: Durable embedding job queued }
        '403': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/overview:
    get:
      operationId: getOverview
      summary: Read the current technical onboarding overview
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      responses:
        '200':
          description: Overview and limitations
          content:
            application/json:
              schema: { $ref: '../contracts/analysis.schema.json#/$defs/Overview' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/overview/inventory:
    get:
      operationId: getOverviewInventory
      summary: Page through every accounted-for snapshot artifact
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
        - name: offset
          in: query
          schema: { type: integer, minimum: 0 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500 }
      responses:
        '200': { description: Artifact status, category, and reason codes }
        '404': { $ref: '#/components/responses/Problem' }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/overview/notes:
    post:
      operationId: addOverviewNote
      summary: Retain a local correction attached to a current overview claim
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [claim_id, note]
              properties:
                claim_id: { type: string }
                note: { type: string, minLength: 1, maxLength: 2000 }
      responses:
        '201': { description: Correction retained without automatic human verification }
        '422': { $ref: '#/components/responses/Problem' }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/overview/notes/{noteId}:
    delete:
      operationId: deleteOverviewNote
      summary: Remove one local correction from the selected snapshot
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
        - name: noteId
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '204': { description: Correction removed }
        '404': { $ref: '#/components/responses/Problem' }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/analysis:
    get:
      operationId: analysisStatus
      summary: Latest synthesis job state for this snapshot
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      responses:
        '200': { description: Latest job ID and state, or not_started }
        '404': { $ref: '#/components/responses/Problem' }
    post:
      operationId: queueAnalysis
      summary: Analyze up to eight uncached source units and publish assessed overview and OKF
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                cloud_opt_in:
                  type: boolean
                  default: false
                  description: Explicitly saves cloud permission for this repository when installation policy allows it.
      responses:
        '202':
          description: Analysis queued; repeat after completion to resume remaining source units
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobAccepted' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/questions:
    get:
      operationId: getConversation
      summary: Most recent thirty answers in this snapshot, oldest first
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      responses:
        '200': { description: Object with answers array }
        '404': { $ref: '#/components/responses/Problem' }
    post:
      operationId: askQuestion
      summary: Retrieve OKF and source, explore bounded relationships, compose and assess a technical answer
      description: Synchronous request with a seven-minute timeout. Mutations require the local session cookie and X-CSRF-Token.
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [question]
              properties:
                question: { type: string, minLength: 1, maxLength: 2000, description: Also limited to 2000 UTF-8 bytes by the server. }
                parent_id: { type: string, description: Previous answer ID from this snapshot, or empty for a new question. }
      responses:
        '200':
          description: Persisted assessed claims, source citations, retrieved OKF concepts, static diagram links and explicit gaps
          content:
            application/json:
              schema:
                $ref: '../contracts/answer.schema.json'
        '404': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/evidence/{evidenceId}:
    get:
      operationId: getEvidence
      summary: Resolve an authorized citation to sanitized source lines
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
        - name: evidenceId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Sanitized evidence; never raw excluded secret content
          content:
            application/json:
              schema: { $ref: '../contracts/evidence.schema.json#/$defs/Evidence' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/knowledge/export:
    get:
      operationId: exportKnowledge
      summary: Download a validated, source-backed OKF v0.2 ZIP bundle
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      responses:
        '200':
          description: Sanitized OKF bundle
          content:
            application/zip:
              schema: { type: string, format: binary }
        '403': { $ref: '#/components/responses/Problem' }
  /v1/jobs/{jobId}:
    get:
      operationId: getJob
      summary: Read progress, partial failures, and result ID
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Job state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '403': { $ref: '#/components/responses/Problem' }
  /v1/jobs/{jobId}/events:
    get:
      operationId: streamJobEvents
      summary: Stream progress for an authorized job
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Server-sent events; payloads contain counts and state, not code or secrets
          content:
            text/event-stream:
              schema: { type: string }
        '403': { $ref: '#/components/responses/Problem' }
components:
  parameters:
    RepositoryId:
      name: repositoryId
      in: path
      required: true
      schema: { type: string, minLength: 1 }
    SnapshotId:
      name: snapshotId
      in: path
      required: true
      schema: { type: string, minLength: 1 }
    JobId:
      name: jobId
      in: path
      required: true
      schema: { type: string, minLength: 1 }
  schemas:
    JobAccepted:
      type: object
      additionalProperties: false
      required: [job_id, state]
      properties:
        job_id: { type: string }
        state: { const: queued }
        repository_id: { type: string }
        snapshot_id: { type: string }
    Job:
      type: object
      additionalProperties: false
      required: [job_id, kind, state]
      properties:
        job_id: { type: string }
        kind: { enum: [ingest_snapshot, embed_snapshot] }
        state: { type: string, description: River job state }
    Problem:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        request_id: { type: string }
  responses:
    Problem:
      description: Safe error without raw input, code, path, or credential value
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Problem' }